backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
sqlalchemy==2.0.22
alembic==1.12.0
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# AI and ML dependencies
openai==1.2.0
numpy==1.26.0
scipy==1.11.3
librosa==0.10.1
soundfile==0.12.1
tensorflow==2.14.0
transformers==4.34.0

# Testing
pytest==7.4.2
httpx==0.25.0

# Utilities
pydub==0.25.1
arabic-reshaper==3.0.0
python-bidi==0.4.2
